+++ x/marker/types/msg.go
@@ -501,12 +501,16 @@ func (msg MsgTransferRequest) GetSigners() []sdk.AccAddress {
 	if err != nil {
 		panic(err)
 	}
-	sourceAddr, err := sdk.AccAddressFromBech32(msg.FromAddress)
-	if err != nil {
-		panic(err)
-	}
+	// NOTE: removed to support smart contract facilitated transfer of RESTRICTED_COIN marker coins.
+	// When SDK 0.43 authz support is available this constraint will be restored using grant based authorizations
+	// for transfer of funds.  See https://github.com/provenance-io/provenance/issues/246
 
-	return []sdk.AccAddress{adminAddr, sourceAddr}
+	// sourceAddr, err := sdk.AccAddressFromBech32(msg.FromAddress)
+	// if err != nil {
+	// 	panic(err)
+	// }
+
+	return []sdk.AccAddress{adminAddr} // sourceAddr}
 }
 
 // NewSetDenomMetadataRequest  creates a new marker in a proposed state with a given total supply a denomination